Move shared code for repositories into base repository.
